+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Fixed
+- Custom configuration properties were incorrectly added to batch properties
+
 ## [5.0.2] - 2022-01-31
 ### Updated
 - Allow installing with older appium and selenium on python>=3.7